refactor: move state modules into lib
