+++ src/lib/project-store.test.ts
